move instialization to the decoder_trainer
