+++ tests/datasets/string_reverse_test.py
@@ -9,7 +9,7 @@ from trainite.datasets.string_reverse import (
 )
 from trainite.datasets.transformed import TransformedDataset
 from trainite.preprocessors.char_tokenizer import CharTokenizer
-from trainite.shared.main import build_dataset
+from trainite.trainers.decoder_trainer import build_dataset
 from trainite.shared.utils import get_target
